+++ Structurizr.AdrTools.Tests/TestFailedException.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace Structurizr.AdrTools.Tests
-{
-    public class TestFailedException : Exception
-    {
-        
-    }
-}